Use the same string serializer
